+++ modules/AWSPowerShell/Cmdlets/AmazonCloudWatch/Basic/Get-CWDashboard-Cmdlet.cs
@@ -1,0 +1,149 @@
+/*******************************************************************************
+ *  Copyright 2012-2015 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *  Licensed under the Apache License, Version 2.0 (the "License"). You may not use
+ *  this file except in compliance with the License. A copy of the License is located at
+ *
+ *  http://aws.amazon.com/apache2.0
+ *
+ *  or in the "license" file accompanying this file.
+ *  This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+ *  CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ *  specific language governing permissions and limitations under the License.
+ * *****************************************************************************
+ *
+ *  AWS Tools for Windows (TM) PowerShell (TM)
+ *
+ */
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Management.Automation;
+using System.Text;
+using Amazon.PowerShell.Common;
+using Amazon.Runtime;
+using Amazon.CloudWatch;
+using Amazon.CloudWatch.Model;
+
+namespace Amazon.PowerShell.Cmdlets.CW
+{
+    /// <summary>
+    /// Displays the details of the dashboard that you specify.
+    /// 
+    ///  
+    /// <para>
+    /// To copy an existing dashboard, use <code>GetDashboard</code>, and then use the data
+    /// returned within <code>DashboardBody</code> as the template for the new dashboard when
+    /// you call <code>PutDashboard</code> to create the copy.
+    /// </para>
+    /// </summary>
+    [Cmdlet("Get", "CWDashboard")]
+    [OutputType("Amazon.CloudWatch.Model.GetDashboardResponse")]
+    [AWSCmdlet("Invokes the GetDashboard operation against Amazon CloudWatch.", Operation = new[] {"GetDashboard"})]
+    [AWSCmdletOutput("Amazon.CloudWatch.Model.GetDashboardResponse",
+        "This cmdlet returns a Amazon.CloudWatch.Model.GetDashboardResponse object containing multiple properties. The object can also be referenced from properties attached to the cmdlet entry in the $AWSHistory stack."
+    )]
+    public partial class GetCWDashboardCmdlet : AmazonCloudWatchClientCmdlet, IExecutor
+    {
+        
+        #region Parameter DashboardName
+        /// <summary>
+        /// <para>
+        /// <para>The name of the dashboard to be described.</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(Position = 0, ValueFromPipeline = true)]
+        public System.String DashboardName { get; set; }
+        #endregion
+        
+        protected override void ProcessRecord()
+        {
+            base.ProcessRecord();
+            
+            var context = new CmdletContext
+            {
+                Region = this.Region,
+                Credentials = this.CurrentCredentials
+            };
+            
+            // allow for manipulation of parameters prior to loading into context
+            PreExecutionContextLoad(context);
+            
+            context.DashboardName = this.DashboardName;
+            
+            // allow further manipulation of loaded context prior to processing
+            PostExecutionContextLoad(context);
+            
+            var output = Execute(context) as CmdletOutput;
+            ProcessOutput(output);
+        }
+        
+        #region IExecutor Members
+        
+        public object Execute(ExecutorContext context)
+        {
+            var cmdletContext = context as CmdletContext;
+            // create request
+            var request = new Amazon.CloudWatch.Model.GetDashboardRequest();
+            
+            if (cmdletContext.DashboardName != null)
+            {
+                request.DashboardName = cmdletContext.DashboardName;
+            }
+            
+            CmdletOutput output;
+            
+            // issue call
+            var client = Client ?? CreateClient(context.Credentials, context.Region);
+            try
+            {
+                var response = CallAWSServiceOperation(client, request);
+                Dictionary<string, object> notes = null;
+                object pipelineOutput = response;
+                output = new CmdletOutput
+                {
+                    PipelineOutput = pipelineOutput,
+                    ServiceResponse = response,
+                    Notes = notes
+                };
+            }
+            catch (Exception e)
+            {
+                output = new CmdletOutput { ErrorResponse = e };
+            }
+            
+            return output;
+        }
+        
+        public ExecutorContext CreateContext()
+        {
+            return new CmdletContext();
+        }
+        
+        #endregion
+        
+        #region AWS Service Operation Call
+        
+        private Amazon.CloudWatch.Model.GetDashboardResponse CallAWSServiceOperation(IAmazonCloudWatch client, Amazon.CloudWatch.Model.GetDashboardRequest request)
+        {
+            Utils.Common.WriteVerboseEndpointMessage(this, client.Config, "Amazon CloudWatch", "GetDashboard");
+            #if DESKTOP
+            return client.GetDashboard(request);
+            #elif CORECLR
+            // todo: handle AggregateException and extract true service exception for rethrow
+            var task = client.GetDashboardAsync(request);
+            return task.Result;
+            #else
+                    #error "Unknown build edition"
+            #endif
+        }
+        
+        #endregion
+        
+        internal class CmdletContext : ExecutorContext
+        {
+            public System.String DashboardName { get; set; }
+        }
+        
+    }
+}