+++ modules/AWSPowerShell/Cmdlets/AmazonCloudWatch/Basic/Get-CWDashboardList-Cmdlet.cs
@@ -1,0 +1,201 @@
+/*******************************************************************************
+ *  Copyright 2012-2015 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *  Licensed under the Apache License, Version 2.0 (the "License"). You may not use
+ *  this file except in compliance with the License. A copy of the License is located at
+ *
+ *  http://aws.amazon.com/apache2.0
+ *
+ *  or in the "license" file accompanying this file.
+ *  This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+ *  CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ *  specific language governing permissions and limitations under the License.
+ * *****************************************************************************
+ *
+ *  AWS Tools for Windows (TM) PowerShell (TM)
+ *
+ */
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Management.Automation;
+using System.Text;
+using Amazon.PowerShell.Common;
+using Amazon.Runtime;
+using Amazon.CloudWatch;
+using Amazon.CloudWatch.Model;
+
+namespace Amazon.PowerShell.Cmdlets.CW
+{
+    /// <summary>
+    /// Returns a list of the dashboards for your account. If you include <code>DashboardNamePrefix</code>,
+    /// only those dashboards with names starting with the prefix are listed. Otherwise, all
+    /// dashboards in your account are listed.<br/><br/>This operation automatically pages all available results to the pipeline - parameters related to iteration are only needed if you want to manually control the paginated output.
+    /// </summary>
+    [Cmdlet("Get", "CWDashboardList")]
+    [OutputType("Amazon.CloudWatch.Model.DashboardEntry")]
+    [AWSCmdlet("Invokes the ListDashboards operation against Amazon CloudWatch.", Operation = new[] {"ListDashboards"})]
+    [AWSCmdletOutput("Amazon.CloudWatch.Model.DashboardEntry",
+        "This cmdlet returns a collection of DashboardEntry objects.",
+        "The service call response (type Amazon.CloudWatch.Model.ListDashboardsResponse) can also be referenced from properties attached to the cmdlet entry in the $AWSHistory stack.",
+        "Additionally, the following properties are added as Note properties to the service response type instance for the cmdlet entry in the $AWSHistory stack: NextToken (type System.String)"
+    )]
+    public partial class GetCWDashboardListCmdlet : AmazonCloudWatchClientCmdlet, IExecutor
+    {
+        
+        #region Parameter DashboardNamePrefix
+        /// <summary>
+        /// <para>
+        /// <para>If you specify this parameter, only the dashboards with names starting with the specified
+        /// string are listed. The maximum length is 255, and valid characters are A-Z, a-z, 0-9,
+        /// ".", "-", and "_". </para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(Position = 0, ValueFromPipeline = true)]
+        public System.String DashboardNamePrefix { get; set; }
+        #endregion
+        
+        #region Parameter NextToken
+        /// <summary>
+        /// <para>
+        /// <para>The token returned by a previous call to indicate that there is more data available.</para>
+        /// </para>
+        /// <para>
+        /// <br/><b>Note:</b> This parameter is only used if you are manually controlling output pagination of the service API call.
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter]
+        public System.String NextToken { get; set; }
+        #endregion
+        
+        protected override void ProcessRecord()
+        {
+            base.ProcessRecord();
+            
+            var context = new CmdletContext
+            {
+                Region = this.Region,
+                Credentials = this.CurrentCredentials
+            };
+            
+            // allow for manipulation of parameters prior to loading into context
+            PreExecutionContextLoad(context);
+            
+            context.DashboardNamePrefix = this.DashboardNamePrefix;
+            context.NextToken = this.NextToken;
+            
+            // allow further manipulation of loaded context prior to processing
+            PostExecutionContextLoad(context);
+            
+            var output = Execute(context) as CmdletOutput;
+            ProcessOutput(output);
+        }
+        
+        #region IExecutor Members
+        
+        public object Execute(ExecutorContext context)
+        {
+            var cmdletContext = context as CmdletContext;
+            
+            // create request and set iteration invariants
+            var request = new Amazon.CloudWatch.Model.ListDashboardsRequest();
+            
+            if (cmdletContext.DashboardNamePrefix != null)
+            {
+                request.DashboardNamePrefix = cmdletContext.DashboardNamePrefix;
+            }
+            
+            // Initialize loop variant and commence piping
+            System.String _nextMarker = null;
+            bool _userControllingPaging = false;
+            if (AutoIterationHelpers.HasValue(cmdletContext.NextToken))
+            {
+                _nextMarker = cmdletContext.NextToken;
+                _userControllingPaging = true;
+            }
+            
+            try
+            {
+                do
+                {
+                    request.NextToken = _nextMarker;
+                    
+                    var client = Client ?? CreateClient(context.Credentials, context.Region);
+                    CmdletOutput output;
+                    
+                    try
+                    {
+                        
+                        var response = CallAWSServiceOperation(client, request);
+                        
+                        Dictionary<string, object> notes = null;
+                        object pipelineOutput = response.DashboardEntries;
+                        notes = new Dictionary<string, object>();
+                        notes["NextToken"] = response.NextToken;
+                        output = new CmdletOutput
+                        {
+                            PipelineOutput = pipelineOutput,
+                            ServiceResponse = response,
+                            Notes = notes
+                        };
+                        if (_userControllingPaging)
+                        {
+                            int _receivedThisCall = response.DashboardEntries.Count;
+                            WriteProgressRecord("Retrieving", string.Format("Retrieved {0} records starting from marker '{1}'", _receivedThisCall, request.NextToken));
+                        }
+                        
+                        _nextMarker = response.NextToken;
+                    }
+                    catch (Exception e)
+                    {
+                        output = new CmdletOutput { ErrorResponse = e };
+                    }
+                    
+                    ProcessOutput(output);
+                    
+                } while (AutoIterationHelpers.HasValue(_nextMarker));
+            }
+            finally
+            {
+                if (_userControllingPaging)
+                {
+                    WriteProgressCompleteRecord("Retrieving", "Retrieved records");
+                }
+            }
+            
+            return null;
+        }
+        
+        public ExecutorContext CreateContext()
+        {
+            return new CmdletContext();
+        }
+        
+        #endregion
+        
+        #region AWS Service Operation Call
+        
+        private Amazon.CloudWatch.Model.ListDashboardsResponse CallAWSServiceOperation(IAmazonCloudWatch client, Amazon.CloudWatch.Model.ListDashboardsRequest request)
+        {
+            Utils.Common.WriteVerboseEndpointMessage(this, client.Config, "Amazon CloudWatch", "ListDashboards");
+            #if DESKTOP
+            return client.ListDashboards(request);
+            #elif CORECLR
+            // todo: handle AggregateException and extract true service exception for rethrow
+            var task = client.ListDashboardsAsync(request);
+            return task.Result;
+            #else
+                    #error "Unknown build edition"
+            #endif
+        }
+        
+        #endregion
+        
+        internal class CmdletContext : ExecutorContext
+        {
+            public System.String DashboardNamePrefix { get; set; }
+            public System.String NextToken { get; set; }
+        }
+        
+    }
+}